% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{bake.step_pretrained_text_embedding}
\alias{bake.step_pretrained_text_embedding}
\title{Bake method for step_pretrained_text_embedding}
\usage{
bake.step_pretrained_text_embedding(object, new_data, ...)
}
\arguments{
\item{object}{object}

\item{new_data}{new data to apply transformations}

\item{...}{One or more selector functions to choose variables.}
}
\description{
Bake method for step_pretrained_text_embedding
}
