% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{hub_add_signature}
\alias{hub_add_signature}
\title{Adds a signature to the module definition.}
\usage{
hub_add_signature(name = NULL, inputs = NULL, outputs = NULL)
}
\arguments{
\item{name}{Signature name as a string. If omitted, it is interpreted as
'default' and is the signature used when `Module.__call__` signature is
not specified.}

\item{inputs}{A list from input name to Tensor or SparseTensor to feed when
applying the signature. If a single tensor is passed, it is interpreted as a
list with a single 'default' entry.}

\item{outputs}{A list from output name to Tensor or SparseTensor to return from
applying the signature. If a single tensor is passed, it is interpreted as a
list with a single 'default' entry.}
}
\description{
Adds a signature to the module definition.
}
\note{
This must be called within a module_fn that is defining a Module.
}
