% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{hub_create_module_spec}
\alias{hub_create_module_spec}
\alias{hub_create_module_spec_from_savedmodel}
\title{Creates a ModuleSpec from a function that builds the module's graph.}
\usage{
hub_create_module_spec(module_fn, tags_and_args = NULL,
  drop_collections = NULL)

hub_create_module_spec_from_savedmodel(saved_model_path,
  drop_collections = NULL)
}
\arguments{
\item{module_fn}{a function to build a graph for the Module.}

\item{tags_and_args}{Optional list of tuples (tags, kwargs) of tags and keyword
args used to define graph variants. If omitted, it is interpreted as `[(set(), {})]`,
meaning module_fn is called once with no args.}

\item{saved_model_path}{Directory with the SavedModel to use.}

\item{drop_collections:}{list of collection to drop.}
}
\description{
The module_fn is called on a new graph (not the current one) to build the graph
of the module and define its signatures via `hub_add_signature()`.
}
\note{
In anticipation of future TF-versions, `module_fn` is called on a graph
  that uses resource variables by default. If you want old-style variables then
  you can use with `tf$variable_scope("", use_resource=False)` in `module_fn`.
}
\seealso{
hub_add_signature
}
