% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{hub_image_embedding_column}
\alias{hub_image_embedding_column}
\title{Module to construct a dense 1-D representation from the pixels of images.}
\usage{
hub_image_embedding_column(key, module_spec)
}
\arguments{
\item{key}{A string or
\href{feature_column}{https://tensorflow.rstudio.com/tfestimators/articles/feature_columns.html}
 identifying the input image data.}

\item{module_spec}{Module spec loaded using [hub_load_module_spec()] or string
describing the location of a module. There are several supported path encoding
schemes: a) URL location specifying an archived module (e.g.
http://domain/module.tgz) b) Any filesystem location of a module directory (e.g.
/module_dir for a local filesystem). All filesystems implementations provided by
Tensorflow are supported.}
}
\description{
Module to construct a dense 1-D representation from the pixels of images.
}
\details{
This feature column can be used on images, represented as float32 tensors of RGB pixel
data in the range [0,1]. This can be read from a
\href{column_numeric()}{https://tensorflow.rstudio.com/tfestimators/reference/column_numeric.html}
if the input data happens to have decoded images, all with the same shape [height,
width, 3]. More commonly, the
\href{input_fn}{https://tensorflow.rstudio.com/tfestimators/articles/input_functions.html}
will have code to explicitly decode images, resize them (possibly after performing data
augmentation such as random crops etc.), and provide a batch of shape [batch_size,
height, width, 3].
}
