% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{hub_load_module_spec}
\alias{hub_load_module_spec}
\title{Loads a module specification from the filesystem or a URL}
\usage{
hub_load_module_spec(path)
}
\arguments{
\item{path}{String describing the location of a module. There are several supported
path encoding schemes: a) URL location specifying an archived module (e.g.
http://domain/module.tgz) b) Any filesystem location of a module directory (e.g.
/module_dir for a local filesystem). All filesystems implementations provided by
Tensorflow are supported.}
}
\description{
Loads a module specification from the filesystem or a URL
}
