% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{hub_module}
\alias{hub_module}
\title{Re-usable TensorFlow module}
\usage{
hub_module(module_spec, trainable = FALSE, name = "module",
  tags = NULL)
}
\arguments{
\item{module_spec}{Module spec loaded using [hub_load_module_spec()] or string
describing the location of a module. There are several supported path encoding
schemes: a) URL location specifying an archived module (e.g.
http://domain/module.tgz) b) Any filesystem location of a module directory (e.g.
/module_dir for a local filesystem). All filesystems implementations provided by
Tensorflow are supported.}

\item{trainable}{Whether the Module is trainable. If `FALSE`, no variables are added to
TRAINABLE_VARIABLES collection, and no tensors are added to REGULARIZATION_LOSSES
collection.}

\item{name}{A string, the variable scope name under which to create the Module. It will
be uniquified and the equivalent name scope must be unused.}

\item{tags}{A set of strings specifying the graph variant to use.}
}
\description{
A Module represents a part of a TensorFlow graph that can be exported to disk (based on
the SavedModel format) and later re-loaded. A Module has a defined interface that
allows it to be used in a replaceable way, with little or no knowledge of its internals
and its serialization format
}
