% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{hub_text_embedding_column}
\alias{hub_text_embedding_column}
\title{Module to construct a dense representation from a text feature.}
\usage{
hub_text_embedding_column(key, module_spec, trainable = FALSE)
}
\arguments{
\item{key}{A string or
\href{feature_column}{https://tensorflow.rstudio.com/tfestimators/articles/feature_columns.html}
 identifying the text feature.}
}
\description{
This feature column can be used on an input feature whose values are strings of
arbitrary size.
}
