% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_tfhub}
\alias{install_tfhub}
\title{Install TensorFlow Hub}
\usage{
install_tfhub(version = "release", ..., restart_session = TRUE)
}
\arguments{
\item{version}{version of TensorFlow Hub to be installed.}

\item{...}{other arguments passed to [reticulate::py_install()].}

\item{restart_session}{Restart R session after installing (note this will
only occur within RStudio).}
}
\description{
This function is used to install the TensorFlow Hub python module.
}
