% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{prep.step_pretrained_text_embedding}
\alias{prep.step_pretrained_text_embedding}
\title{Prep method for step_pretrained_text_embedding}
\usage{
prep.step_pretrained_text_embedding(x, training, info = NULL, ...)
}
\arguments{
\item{x}{object}

\item{training}{wether or not it's training}

\item{info}{variables state}

\item{...}{One or more selector functions to choose variables.}
}
\description{
Prep method for step_pretrained_text_embedding
}
