% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{hub_image_embedding_column}
\alias{hub_image_embedding_column}
\title{Module to construct a dense 1-D representation from the pixels of images.}
\usage{
hub_image_embedding_column(key, module_spec)
}
\arguments{
\item{key}{A string or [feature_column](https://tensorflow.rstudio.com/tfestimators/articles/feature_columns.html)
identifying the text feature.}

\item{module_spec}{A string handle or a _ModuleSpec identifying the module.}
}
\description{
Module to construct a dense 1-D representation from the pixels of images.
}
\details{
This feature column can be used on images, represented as float32 tensors of RGB pixel
data in the range [0,1].
}
